added class method to Order model for getting cart infomration and updated route to use this method
